requirements.txt
alembic==1.13.0
aniso8601==9.0.1
blinker==1.7.0
Cerberus==1.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
configparser==6.0.0
cryptography==42.0.5
Deprecated==1.2.14
elastic-transport==8.11.0
elasticsearch==8.11.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Headers==1.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
idna==3.6
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
pycparser==2.22
PyJWT==2.8.0
pykakasi==2.2.1
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
uWSGI==2.0.23
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.1
